hardware: fix outline, replace phantom pad with real one
There was a phantom pad that was created with an external graphic.  It's
unselectable in Kicad, and is weird.  This pad exposes the soldermask on
the captouch buttons.

Recreate it as a real item.

Additionally, fix the graphic outline.

Additionally, move everything up to the corner, to fix drill files later
on.

Signed-off-by: Sean Cross <sean@xobs.io>