pcb: increase copper going to 3.3V pad
Reroute one captouch trace and use two vias to tap into the 3.3V rail on
the backside.  This gives us more reliable 3.3V on pin 2.  Additionally,
we now have two vias going to the backside 3.3V rail, which should
improve the power flow to the other half of the board.

Signed-off-by: Sean Cross <sean@xobs.io>